Exam mode: move import-progress enums out of the dto package, fix spec mock types
CI caught two issues the local checks missed:
- ExamCodeStyleArchitectureTest requires classes in the dto package to be named *DTO.
  The progress enums are not DTOs, so move ExamImportProgressState and
  ExerciseImportStatus to the exam.domain package (where the other exam enums live).
- The spec typecheck (compile:tests / tsc) rejected two "wrong user input" test mocks
  that still returned the pre-DTO body shape. Vitest runs via esbuild and skips type
  checking, so these passed locally; align the mock bodies with the new return types.
